requirements.txt
streamlit==1.37.1
pdfplumber==0.11.4
python-docx==1.1.2
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.2
sentence-transformers==3.0.1
torch>=2.2.0
